+++ ElvUI_MerathilisUI/Core/Changelog/6.49.lua
@@ -1,8 +1,13 @@
 local MER = unpack(ElvUI_MerathilisUI)
 
 MER.Changelog[649] = {
-	RELEASE_DATE = "TBD",
-	FIXES = {},
+	RELEASE_DATE = "22.12.2024",
+	FIXES = {
+		"[World Map]: Updated the Navigation Bar look on the top",
+		"[World Map]: Updated the Map Reveal Data",
+	},
 	NEW = {},
-	IMPROVEMENTS = {},
+	IMPROVEMENTS = {
+		"[AutoButtons]: Add more reward boxes",
+	},
 }